copied netcdf-4.3 DLLs from python's site-packages to our third_party_lib using the dlls from UCARR site was causing DLL import errors since the DLLs packaged with python netCDF4 module are different. For now, just copied these over to PyGnome 3rd party lib so things work, but still looking for better solution.

+++ py_gnome/setup.py
@@ -156,10 +156,25 @@ if sys.platform is "darwin" or "win32":
                               sys.platform, architecture)
     netcdf_libs = os.path.join(netcdf_base, 'lib')
     netcdf_inc = os.path.join(netcdf_base, 'include')
-    win_dlls = os.path.join(netcdf_base, 'bin')
 
     if sys.platform == 'win32':
         # also copy the netcdf *.dlls to cy_gnome directory
+        # On windows the dlls have the same names for those used by python's
+        # netCDF4 module and PyGnome modules. For PyGnome, we had the latest
+        # netcdf dlls from UCARR site but this was giving DLL import errors.
+        # The netcdf dlls that come with Python's netCDF4 module (C. Gohlke's site)
+        # were different from the netcdf4 DLLs we got from UCARR.
+        # For now, third_party_lib contains the DLLs installed in site-packages
+        # from C. Gohlke's site.
+        #
+        # Alternatively, we could also look for python netCDF4 package and copy DLLs from site-packages.
+        # This way the DLLs used and loaded by PyGnome are the same as the DLL used and
+        # expected by netCDF4. PyGnome loads the DLL with cy_basic_types.pyd and it also
+        # imports netCDF4 when netcdf_outputters module is imported - this was causing the
+        # previous conflict. The DLL loaded in memory should be consistent - that's the best
+        # understanding of current issue!
+        # STILL WORKING ON A MORE PERMANENT SOLUTION
+        win_dlls = os.path.join(netcdf_base, 'bin')
         cwd = os.getcwd()
         dlls_path = os.path.join(cwd, SETUP_PATH, win_dlls)
 
@@ -169,8 +184,9 @@ if sys.platform is "darwin" or "win32":
             if sys.argv[1] == 'cleanall' or sys.argv[1] == 'clean':
                 (x_, dll_name) = os.path.split(dll)
                 rm_dll = os.path.join(dlls_dst, dll_name)
-                os.remove(rm_dll)
-                print "deleted: " + rm_dll
+                if os.path.exists(rm_dll):
+                    os.remove(rm_dll)
+                    print "deleted: " + rm_dll
             else:
                 print "copy: " + dll + " to: " + dlls_dst
                 shutil.copy(dll, dlls_dst)
